cleanup: Make a pass cleaning up warnings
Cleanup a lot of warnings of signed/unsigned comparisons,
calling return on void methods etc.

+++ src/lib/fy-diag.c
@@ -800,14 +800,14 @@ void fy_diag_error_atom_display(struct fy_diag *diag, enum fy_error_type type, s
 				}
 			}
 			display = rowbuf;
-			display_len = rbs - rowbuf;
+			display_len = (int)(rbs - rowbuf);
 
 			tilde_start = content_start_col - line_shift;
 			tilde_width = content_width;
 			if (tilde_start + tilde_width > cols)
 				tilde_width = cols - tilde_start;
 			if ((size_t)tilde_width >= rowbufsz)
-				tilde_width = rowbufsz - 1;	/* guard */
+				tilde_width = (int)(rowbufsz - 1);	/* guard */
 			tilde_width_m1 = tilde_width > 0 ? (tilde_width - 1) : 0;
 
 			/* output the line */
